fix: Extract chrome storage functions from api
